Fix ACE-Step float16 on L4 and tighten RunPod deploy to flashboot + L4 only.
Defer model warmup remains; bed_loop now reports ml source correctly when AI succeeds.

+++ apps/worker/pipeline/bed_ai.py
@@ -257,7 +257,10 @@ def _load_acestep():
     from diffusers import AceStepPipeline
 
     model_id = os.environ.get("ACESTEP_MODEL", "ACE-Step/acestep-v15-xl-turbo-diffusers")
-    pipe = AceStepPipeline.from_pretrained(model_id, torch_dtype=torch.bfloat16)
+    pipe = AceStepPipeline.from_pretrained(
+        model_id,
+        torch_dtype=torch.float16,
+    )
     return pipe.to("cuda")
